==> finetuning_code/0_prepare_sets_llm.ipynb <==
import pandas as pd
from sklearn.model_selection import train_test_split
import os
import sys
os.chdir("../../../../../../../../")
os.chdir("home/user/mnt/degelin/thesis")
sys.path.append("tune_sets/")

def split_dataset(filename, train_size=0.7, val_size=0.1, test_size=0.2, random_state=42):
    base_path = 'tune_sets/llm/'
    data = pd.read_csv(base_path + filename)
    
    # First split: separate out the test set
    train_val_data, test_data = train_test_split(data, test_size=test_size, random_state=random_state, stratify=data['label'])
    
    # Second split: separate train and validation from the remaining data
    train_data, val_data = train_test_split(train_val_data, 
                                            test_size=val_size/(train_size+val_size), 
                                            random_state=random_state, 
                                            stratify=train_val_data['label'])
    
    # Print the value counts for each split to ensure they are balanced
    print("Training set: \n", train_data['label'].value_counts())
    print("Validation set: \n", val_data['label'].value_counts())
    print("Test set: \n", test_data['label'].value_counts())
    
    # Save the split datasets
    train_filename = base_path + filename.replace('.csv', '_train_part.csv')
    val_filename = base_path + filename.replace('.csv', '_val_part.csv')
    test_filename = base_path + filename.replace('.csv', '_test_part.csv')
    train_data.to_csv(train_filename, index=False)
    val_data.to_csv(val_filename, index=False)
    test_data.to_csv(test_filename, index=False)

    return train_data, val_data, test_data

filename = 'reality_withoutNames.csv'
model = SentenceTransformer(model_name)
tokenizer = model.tokenizer# Replace with your actual filename
split_dataset(filename)
