==> finetuning_code/1_load_libraries.ipynb <==
import os
os.chdir("../../../")
file_path = "biased-rulers/requirements.txt"
os.chdir("../../../../../../../../")
os.chdir("home/user/mnt/degelin/thesis")
# Read and print the contents of the file
with open(file_path, "r") as file:
    print(file.read())
!pip install -r biased-rulers/requirements.txt
!pip install sentence-transformers==2.2.2
