==> finetuning_code/2_intrinsic_bias.ipynb <==
!pip install pytest
from transformers import AutoTokenizer, AutoModelForMaskedLM, AutoModel
import pytest
import numpy as np
from pytest import approx
import sys
import os
os.chdir("../../../../../../../../")
os.chdir("home/user/mnt/degelin/thesis")
sys.path.append("biased-rulers/")
from biased_rulers.metrics import disco
from biased_rulers.metrics import lpbs
from biased_rulers.metrics import seat

def test_disco_bert_base_names(model_name):
    os.chdir("biased-rulers/biased_rulers/metrics")
    #model = SentenceTransformer(model_name)
    model_type = model_name
    tokenizer = AutoTokenizer.from_pretrained(model_type)
    model = AutoModelForMaskedLM.from_pretrained(model_type)
    print(f"Loaded {model_type}")
    score = disco.disco_test(tokenizer, model)
    os.chdir("../../..")
    # Score for the bert-base names test (taken from https://arxiv.org/pdf/2010.06032.pdf)
    assert score == approx(0.8, abs=1e-1)

def test_lpbs_bert_base_names(model_name):
    os.chdir("biased-rulers/biased_rulers/metrics")
    model_type = model_name
    tokenizer = AutoTokenizer.from_pretrained(model_type)
    model = AutoModelForMaskedLM.from_pretrained(model_type)
    print(f"Loaded {model_type}")
    score = lpbs.lpbs_test(tokenizer, model)
    os.chdir("../../..")
    # Score for the bert-base names test (taken from https://arxiv.org/pdf/2010.06032.pdf)
    # assert score == approx(0.8, abs=1e-1)
    print("lpbs score is")
    print(score)

def test_seat(model_name):
    os.chdir("biased-rulers/biased_rulers/metrics")
    model_type = model_name
    tokenizer = AutoTokenizer.from_pretrained(model_type)
    model = AutoModelForMaskedLM.from_pretrained(model_type)
    print(f"Loaded {model_type}")
    attribute_template = "This is the _."
    target_template = "This is the _."
    results = seat.seat_test(attribute_template, target_template, tokenizer, model)
    score = np.fromiter(results.values(), dtype=float).mean()
    os.chdir("../../..")
    print("original seat score is")
    print(score)
    assert score == approx(0.4365, abs=1e-3)

def test_seat_tan_et_al(model_name):
    os.chdir("biased-rulers/biased_rulers/metrics")
    model_type = model_name
    tokenizer = AutoTokenizer.from_pretrained(model_type)
    model = AutoModelForMaskedLM.from_pretrained(model_type)
    print(f"Loaded {model_type}")
    attribute_template = "This is the _."
    target_template = "This is the _."
    results = seat.tan_et_al_test(attribute_template, target_template, tokenizer, model)
    score = np.fromiter(results.values(), dtype=float).mean()
    os.chdir("../../..")
    print("seat_tan_et_al score is")
    print(score)


def test_seat_lauscher_et_al(model_name):
    os.chdir("biased-rulers/biased_rulers/metrics")
    model_type = model_name
    tokenizer = AutoTokenizer.from_pretrained(model_type)
    model = AutoModelForMaskedLM.from_pretrained(model_type)
    print(f"Loaded {model_type}")
    attribute_template = "This is the _."
    target_template = "This is the _."
    results = seat.lauscher_et_al_test(attribute_template, target_template, tokenizer, model)
    score = np.fromiter(results.values(), dtype=float).mean()
    os.chdir("../../..")
    print("lauscher_et_al seat score is")
    print(score)

test_disco_bert_base_names('bert-base-uncased')

test_lpbs_bert_base_names('bert-base-uncased')

test_seat("bert-base-multilingual-uncased")

test_seat_tan_et_al("bert-base-multilingual-uncased")

test_seat_lauscher_et_al("bert-base-multilingual-uncased")
